@@ -13,39 +13,45 @@
         }
     ],
     "require": {
-        "php": "^7.4",
+        "php": "^8.3",
         "ext-PDO": "*",
         "ext-bz2": "*",
         "ext-json": "*",
-        "ext-zip": "*",
         "ext-sodium": "*",
-        "doctrine/annotations": "^1.13.2",
-        "egulias/email-validator": "3.1.2",
-        "google/apiclient": "^v2.12.1",
-        "google/cloud": "^v0.175.0",
-        "google/protobuf": "^v3.19.4",
-        "google/recaptcha": "^1.2.4",
-        "grpc/grpc": "^1.42.0",
-        "guzzlehttp/guzzle": "^7.4.1",
-        "jolicode/slack-php-api": "^v4.4.1",
-        "kreait/firebase-php": "^6.1.0",
-        "lcobucci/jwt": "^4.1.5",
-        "nesbot/carbon": "^2.56.0",
-        "nyholm/psr7": "^1.5.0",
-        "php-di/php-di": "^6.3.5",
-        "phpmailer/phpmailer": "^v6.5.3",
-        "ramsey/uuid": "^4.2.3",
-        "robmorgan/phinx": "^0.12.10",
-        "sendgrid/sendgrid": "^7.11.4",
+        "ext-zip": "*",
+        "doctrine/annotations": "^1.14.3",
+        "egulias/email-validator": "4.0.2",
+        "google/apiclient": "^v2.16.0",
+        "google/cloud": "^0.246.0",
+        "google/common-protos": "^4.6",
+        "google/protobuf": "^v4.27.0",
+        "google/recaptcha": "^1.3.0",
+        "grpc/grpc": "^1.57.0",
+        "guzzlehttp/guzzle": "^7.8.1",
+        "jolicode/slack-php-api": "^v4.7.0",
+        "kreait/firebase-php": "^7.12.0",
+        "lcobucci/jwt": "^5.3.0",
+        "nesbot/carbon": "^3.4.0",
+        "nyholm/psr7": "^1.8.1",
+        "php-di/php-di": "^7.0.6",
+        "phpmailer/phpmailer": "^v6.9.1",
+        "ramsey/uuid": "^4.7.6",
+        "robmorgan/phinx": "^0.14.0",
+        "sendgrid/sendgrid": "^8.1.2",
         "slim/php-view": "^3.2.0",
-        "slim/psr7": "1.6",
-        "slim/slim": "^4.11.0",
-        "symfony/http-client": "^v5.3.14",
-        "symfony/validator": "^v5.4.3",
-        "zircote/swagger-php": "^4.2.8"
+        "slim/psr7": "1.6.1",
+        "slim/slim": "^4.13.0",
+        "symfony/http-client": "^v7.0.7",
+        "symfony/validator": "^v7.0.7",
+        "zircote/swagger-php": "^4.9.2"
     },
     "autoload": {
-      "psr-4": {"RedCrossQuest\\": "src/"}
+      "psr-4": {
+          "RedCrossQuest\\": "src/"
+      },
+        "exclude-from-classmap": [
+            "vendor/google/cloud/CommonProtos/metadata/Rpc/Context/AttributeContext.php"
+        ]
     },
 
     "config": {
@@ -56,8 +62,8 @@
             "*": "dist"
         },
         "platform": {
-            "php": "7.4",
-            "ext-sodium": "7.4"
+            "php": "8.3",
+            "ext-sodium": "8.3"
         },
         "allow-plugins": {
             "php-http/discovery": true
