@@ -4,7 +4,7 @@
 namespace RedCrossQuest\routes\routesActions\authentication;
 
 
-use DI\Annotation\Inject;
+use DI\Attribute\Inject;
 use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
@@ -33,9 +33,9 @@ class GetUserInfoFromUUIDAction extends Action
    */
   private $queteurDBService;
   /**
-   * @Inject("settings")
    * @var array settings
    */
+  #[Inject("settings")]
   protected $settings;
 
   /**
