@@ -2,7 +2,7 @@
 namespace RedCrossQuest\routes\routesActions\queteurs;
 
 
-use DI\Annotation\Inject;
+use DI\Attribute\Inject;
 use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
@@ -35,9 +35,9 @@ class ApproveQueteurRegistration extends Action
   private $pubSubService;
 
   /**
-   * @Inject("settings")
    * @var array settings
    */
+  #[Inject("settings")]
   protected $settings;
 
   /**
