@@ -6,7 +6,7 @@
 namespace RedCrossQuest\routes\routesActions\queteurs;
 
 
-use DI\Annotation\Inject;
+use DI\Attribute\Inject;
 use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
@@ -38,9 +38,9 @@ class AssociateRegistrationWithExistingQueteur extends Action
   private PubSubService $pubSubService;
 
   /**
-   * @Inject("settings")
    * @var array settings
    */
+  #[Inject("settings")]
   protected array $settings;
 
   /**
