@@ -6,7 +6,7 @@
 namespace RedCrossQuest\routes\routesActions\queteurs;
 
 
-use DI\Annotation\Inject;
+use DI\Attribute\Inject;
 use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
@@ -30,9 +30,9 @@ class CreateQueteur extends Action
    */
   private $emailBusinessService;
   /**
-   * @Inject("settings")
    * @var array settings
    */
+  #[Inject("settings")]
   protected $settings;
 
   /**
