@@ -2,7 +2,7 @@
 namespace RedCrossQuest\routes\routesActions\troncsQueteurs;
 
 
-use DI\Annotation\Inject;
+use DI\Attribute\Inject;
 use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
@@ -49,9 +49,9 @@ class PrepareTroncQueteur extends Action
 
 
   /**
-   * @Inject("settings")
    * @var array settings
    */
+  #[Inject("settings")]
   protected $settings;
 
   /**
