@@ -6,7 +6,7 @@
 namespace RedCrossQuest\routes\routesActions\troncsQueteurs;
 
 
-use DI\Annotation\Inject;
+use DI\Attribute\Inject;
 use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
@@ -32,9 +32,9 @@ class SaveAsAdminOnTroncQueteur extends Action
   private $pubSubService;
 
   /**
-   * @Inject("settings")
    * @var array settings
    */
+  #[Inject("settings")]
   protected $settings;
 
   /**
